Has finished implementation controllers methods POST, PUT and DELETE

+++ src/main/java/com/study/courses/webservice/controller/SubjectController.java
@@ -1,5 +1,7 @@
 package com.study.courses.webservice.controller;
 
+import com.study.courses.webservice.domain.CEFR;
+import com.study.courses.webservice.domain.Language;
 import com.study.courses.webservice.domain.Subject;
 import com.study.courses.webservice.domain.Teacher;
 import com.study.courses.webservice.service.EducationProcessService;
@@ -7,6 +9,7 @@ import com.study.courses.webservice.service.EducationReaderService;
 import lombok.AllArgsConstructor;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
+import org.springframework.hateoas.Link;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,35 +31,38 @@ public class SubjectController {
     @GetMapping()
     public ResponseEntity<CollectionModel<Subject>> findAll() {
         List<Subject> subjects = educationReaderService.findAllSubjects();
+        subjects.forEach(s ->
+                s.add(linkTo(methodOn(SubjectController.class).findSubject(s.getId())).withSelfRel())
+                        .add(linkTo(methodOn(TeacherController.class).find(s.getTeacher().getId())).withRel("teacher")));
+        return ResponseEntity.ok(CollectionModel.of(subjects));
+    }
+
+    @GetMapping("/language/{id}")
+    public ResponseEntity<CollectionModel<Subject>> findAllByLanguage(@PathVariable Long id) {
+        Language language = educationReaderService.findLanguage(id);
+        List<Subject> subjects = educationReaderService.findAllSubjects(language);
         return addLinksSubjects(subjects);
     }
 
-//    @GetMapping("/{idLanguage}")
-//    public ResponseEntity<CollectionModel<Subject>> findAll(@PathVariable Long idLanguage) {
-//        Language language = educationReaderService.findLanguage(idLanguage);
-//        List<Subject> subjects = educationReaderService.findAllSubjects(language);
-//        return addLinksSubjects(subjects);
-//    }
+    @GetMapping("/cefr/{value}")
+    public ResponseEntity<CollectionModel<Subject>> findAllByCEFR(@PathVariable CEFR value) {
+        List<Subject> subjects = educationReaderService.findAllSubjects(value.name());
+        return addLinksSubjects(subjects);
+    }
 
-//    @GetMapping("/{CEFR}")
-//    public ResponseEntity<CollectionModel<Subject>> findAll(@PathVariable String CEFR) {
-//        List<Subject> subjects = educationReaderService.findAllSubjects(CEFR);
-//        return addLinksSubjects(subjects);
-//    }
-
-    @GetMapping("/{idTeach}")
-    public ResponseEntity<CollectionModel<Subject>> findAllByTeacher(@PathVariable Long idTeach) {
-        Teacher teacher = educationReaderService.findTeacher(idTeach);
+    @GetMapping("/teacher/{id}")
+    public ResponseEntity<CollectionModel<Subject>> findAllByTeacher(@PathVariable Long id) {
+        Teacher teacher = educationReaderService.findTeacher(id);
         List<Subject> subjects = educationReaderService.findAllSubjects(teacher);
         return addLinksSubjects(subjects);
     }
 
-//    @GetMapping("/{idLanguage}&{CEFR}")
-//    public ResponseEntity<CollectionModel<Subject>> findAll(@PathVariable Long idLanguage, @PathVariable String CEFR) {
-//        Language Language = educationReaderService.findLanguage(idLanguage);
-//        List<Subject> subjects = educationReaderService.findAllSubjects(Language, CEFR);
-//        return addLinksSubjects(subjects);
-//    }
+    @GetMapping("/language/{id}/cefr/{value}")
+    public ResponseEntity<CollectionModel<Subject>> findAll(@PathVariable Long id, @PathVariable String value) {
+        Language Language = educationReaderService.findLanguage(id);
+        List<Subject> subjects = educationReaderService.findAllSubjects(Language, value);
+        return addLinksSubjects(subjects);
+    }
 
     @GetMapping("/subject/{id}")
     public ResponseEntity<EntityModel<Subject>> findSubject(@PathVariable Long id) {
@@ -74,23 +80,44 @@ public class SubjectController {
 
     @PostMapping("/subject")
     public ResponseEntity<Subject> save(@Valid @RequestBody Subject subject) {
-        return ResponseEntity.ok(educationProcessService.save(subject));
+        Subject currentSubject = educationProcessService.save(subject);
+        if (currentSubject == null) {
+            return ResponseEntity.badRequest().build();
+        } else {
+            Link link = linkTo(methodOn(SubjectController.class).findSubject(currentSubject.getId())).withSelfRel();
+
+            return ResponseEntity.created(link.toUri()).build();
+        }
     }
 
     @PutMapping("/subject")
     public ResponseEntity<Subject> update(@Valid @RequestBody Subject subject) {
-        return ResponseEntity.ok(educationProcessService.save(subject));
+        Subject currentSubject = educationReaderService.findSubject(subject.getId());
+        if (currentSubject == null) {
+            return ResponseEntity.badRequest().build();
+        } else {
+            educationProcessService.save(subject);
+            return ResponseEntity.accepted().build();
+        }
     }
 
     @DeleteMapping("/subject")
-    public void delete(@RequestBody Subject subject) {
-        educationProcessService.delete(subject);
+    public ResponseEntity<Object> delete(@RequestBody Subject subject) {
+        Subject currentSubject = educationReaderService.findSubject(subject.getId());
+        if (currentSubject == null) {
+            return ResponseEntity.badRequest().build();
+        } else {
+            educationProcessService.delete(subject);
+
+            return ResponseEntity.accepted().build();
+        }
     }
 
     private ResponseEntity<CollectionModel<Subject>> addLinksSubjects(List<Subject> subjects) {
         subjects.forEach(s ->
                 s.add(linkTo(methodOn(SubjectController.class).findSubject(s.getId())).withSelfRel())
-                        .add(linkTo(methodOn(TeacherController.class).find(s.getTeacher().getId())).withRel("teacher")));
+                        .add(linkTo(methodOn(TeacherController.class).find(s.getTeacher().getId())).withRel("teacher"))
+                        .add(linkTo(methodOn(SubjectController.class).findAll()).withRel("subjects")));
         return ResponseEntity.ok(CollectionModel.of(subjects));
     }
 }
